+++ crates/cheatnet/src/runtime_extensions/call_to_blockifier_runtime_extension/execution/cairo1_execution.rs
@@ -5,12 +5,13 @@ use crate::runtime_extensions::call_to_blockifier_runtime_extension::execution::
 };
 use crate::runtime_extensions::cheatable_starknet_runtime_extension::CheatableStarknetRuntimeExtension;
 use crate::runtime_extensions::common::get_relocated_vm_trace;
-use blockifier::execution::contract_class::CompiledClassV1;
+use blockifier::execution::contract_class::{CompiledClassV1, TrackedResource};
 use blockifier::execution::entry_point::ExecutableCallEntryPoint;
 use blockifier::execution::entry_point_execution::{
     ExecutionRunnerMode, VmExecutionContext, finalize_execution,
     initialize_execution_context_with_runner_mode, prepare_call_arguments,
 };
+use blockifier::execution::syscalls::vm_syscall_utils::SyscallUsageMap;
 use blockifier::{
     execution::{
         contract_class::EntryPointV1, entry_point::EntryPointExecutionContext,
@@ -26,6 +27,7 @@ use cairo_vm::{
 use runtime::{ExtendedRuntime, StarknetRuntime};
 
 // blockifier/src/execution/cairo1_execution.rs:48 (execute_entry_point_call)
+#[expect(clippy::result_large_err)]
 pub(crate) fn execute_entry_point_call_cairo1(
     call: ExecutableCallEntryPoint,
     compiled_class_v1: &CompiledClassV1,
@@ -93,6 +95,8 @@ pub(crate) fn execute_entry_point_call_cairo1(
     })?;
 
     let trace = get_relocated_vm_trace(&mut runner);
+
+    // Syscall usage here is flat, meaning it only includes syscalls from current call
     let syscall_usage = cheatable_runtime
         .extended_runtime
         .hint_handler
@@ -124,15 +128,22 @@ pub(crate) fn execute_entry_point_call_cairo1(
             .register_error(class_hash, pcs);
     }
 
+    let (syscall_usage_vm_resources, syscall_usage_sierra_gas) = match tracked_resource {
+        TrackedResource::CairoSteps => (syscall_usage, SyscallUsageMap::default()),
+        TrackedResource::SierraGas => (SyscallUsageMap::default(), syscall_usage),
+    };
+
     Ok(CallInfoWithExecutionData {
         call_info,
-        syscall_usage,
+        syscall_usage_vm_resources,
+        syscall_usage_sierra_gas,
         vm_trace: Some(trace),
     })
     // endregion
 }
 
 // crates/blockifier/src/execution/cairo1_execution.rs:236 (run_entry_point)
+#[expect(clippy::result_large_err)]
 pub fn cheatable_run_entry_point(
     runner: &mut CairoRunner,
     hint_processor: &mut dyn HintProcessor,