+++ crates/sncast/tests/e2e/deploy.rs
@@ -224,8 +224,6 @@ async fn test_happy_case_with_constructor_cairo_expression_calldata() {
     assert!(matches!(receipt, Deploy(_)));
 }
 
-// TODO(#3116): Before, this test returned message 'Input too long for arguments').
-// Now, it returns message about transaction execution error.
 #[test]
 fn test_wrong_calldata() {
     let args = vec![
@@ -249,7 +247,7 @@ fn test_wrong_calldata() {
         output,
         indoc! {r"
         Command: deploy
-        Error: Transaction execution error [..]
+        Error: Transaction execution error [..]Input too long for arguments[..]
         "},
     );
 }
@@ -277,8 +275,6 @@ async fn test_contract_not_declared() {
     );
 }
 
-// TODO(#3116): Before, this test returned message containing info that contract is already deployed.
-// Now, it returns message about transaction execution error.
 #[test]
 fn test_contract_already_deployed() {
     let args = vec![
@@ -302,7 +298,7 @@ fn test_contract_already_deployed() {
         output,
         indoc! {r"
         Command: deploy
-        Error: Transaction execution error [..]
+        Error: Transaction execution error [..] contract already deployed at address [..]
         "},
     );
 }